+++ ddk-manager/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Removed
+
+- DLC channel management, persistence, and chain monitoring.
+
 ## [0.7.3](https://github.com/bennyhodl/dlcdevkit/compare/ddk-manager-v0.7.2...ddk-manager-v0.7.3) - 2025-01-17
 
 ### Other